Refactoring of command type
This patch introduces a command variant in the Ciseau main module.

A command can be stop, a no-op, or a movement.

Filebuffer movements are refactored to fit the movement command. This
allows to eliminate a bunch of duplicate management code for view
adjustment and cursor/filebuffer updates.

This is going to greatly simplify the implementation of redo (dot key)
and do n times.