Experiment to support current line highlightning
To make it work I would need to add padding at the filebuffer layer.
However this requires making the filebuffer aware of the length of the
view port, which is tricky because it create friction with the
management of line numbers, long lines, and line breaks.

On one hand it would feel natural to let the composition layer manage
the line numbering, long line wrapping, and line padding

On the other hand, doing view centering require and view adjustement
requires the Filebuffer to know about the view dimensions.

This might require introducing a layer between composition and
filebuffer for view management.