Screen: add annotations for Screen migration
add comments to places in the Ciseau module where previous Screen /
Term writing apis are still used, and prepare for migration to using
the composition buffer.

Expectations are that this should simplify a lot line padding and line
overwrapping.

This should also enable pretty immediately some basic syntax
highlightning, selection highlighnting, and so on.