Change file view printing to use a List
Array of variant of struct was less efficient than a list of struct,
especially since the variant was only used to signifiy the end of the
view.

More importantly I realized that the thing managing the view around the
cursor into file needs to know about line overwrapping.

This pretty much forces that "thing" to also be the Screen for drawing,
since it means it needs to know the size