Move render from Screen to CompositionBuffer
In order to properly support multiple screen, I need to move render to
the single CompositionBuffer which will be used by multiple screens, and
introduce a screen_offset for Screen within a CompositionBuffer.

This also removes the render_buffer from Screen, and adds it along with
the CompositionBuffer to the editor struct. The key insight is that the
editor, CompositionBuffer and render_buffer structs are all singletons.

This patch is a first step in that direction, with some nice cleanups as
a consequence.

The next step is to use size and offset fields in Screen to cleanup a
some vec correcting arguments around.