Introduce signature for Screen
Also removes the FrameBufferImpl functor and simply define the
FrameBuffer module with a direct type constraint to Bytevector.

This is ok to not parametrize Bytevector impl inside FrameBuffer because
I am actually not planning to have any other impl, and the only goal is
to regroup all module signatures at the tope of the file as abstract things
without their implementations. Parametrization of modules is a non-goal
at the moment.