Enable screen offset in put_{block, text, line}
That was easy after the previous refactoring.

Playing with screen size and offset a little bit, this works and a file
can now be displayed in smaller windows wihtin the terminal.

However adding an offset breaks the cursor position, whose relative
position correction is still completely hardcoded.

So now it should be straightforward to support dynamically resizing the
editor:
  1) first let's introduce a function that recreate the editor from the
  term size
  2) then let's bind it to a key
  3) finally, let's handle SIGWINCH to autotrigger it