Hook Tileset into the editor
Unfortunately I had to break the Move and View commands for now to get
it done.

To restore Views and Moves, I need to decide where they get applied.

The cleanest might be to expose a focused_view get/set function pair
in Tileset and use these to manipulate the fileview from the editor.