Eliminate remaining usage of Fileview.view_diff
Scrolling down still crashes in fileviews whose screen is smaller than
the term height.

Also, if the cursor is positionned low enough in the screen and the
tileset layout is changed so that that screen is reduced in height, that
can crash the program.

Is the issue caused by cursor positioning ?