Ground work for tet rendring v3
Another attempt to text rendering is needed !
Some issues with the current system:
 - too many Block.t instances because of tying together colors and text
 - linebreaking is complicated, an dprobably buggy
 - does not cover very well all case, like put_line and selection
   highlightning

The v3 attempt will go back to the previous approach of putting color
sections and text separately. However everything is now put together in
a single Textview. This is required for correct text breaking, and was
the initial insight of v2: to know the actual onscreen offset, one has
to know the exact text that is put on screen before.

This time around, the render logic will follow approximately the
following flow:
 - obtain text for current fileview offset and current screen height
 - create the y offset map from text
 - use the y offset map to output text line per line
 - also use the y offset map to set all colors segments
    - in the future this should be able to be nicely updatable with
      blending modes
 - compute cursor positions