Fix a stupid Tileset bug
The bug which was causing leakage between screens across tiles and
crashing the program in rows layout was due to the following mismatch:

init screen was expecting a pair of offset v2 and screen size v2,
whereas make_screen was producing rectangles with absolute coordinates.
There was therefore a mismatch between the meeing of the 'bottomright'
field in the Rect.rect struct.

This patch fixes this and adds a bunch of asserts for catching issues
earlier.

Remaining bugs:
 = crash when moving the cursor in the lower part of a screen, then changing the
 layout so that this lower part becomes out of the tile
   - the fix should be to readjust the cursors on screen layout changes
 - the tile size computation leaves some blank in Rows and RowMajor mode
   due to integer division having a non-zero remainder.