Complete fileview drawing migration to frame + text
This patch completes the migration of how a fileview is put on the
screen and move from a single put_text to two, one for the header and
the border, and a second one for the text.

This helps with layout and coloring. This also helps with managing
offsets, clipping, and overflow.

Additionally, Colorblock is changed to use any Area.