Another iteration around put_text ...
There was an issue with previous draw_text / draw_frame was that I did
not foresee: the linebreaking of text also affect the line numbering !

That left me with a few bad solutions like getting back the offset map
from Screen.put_text to apply on line numbering.

Instead this patch is an attempt at moving the linebreaking logic one
layer up directly into Fileview. That way I can iteratively populate the
line number array and the text array.

It seems to work, but there are some bugs left.