Fix Overflows cursor position bug
The bug was quite straightforward actually: in the input of
fill_linesinfo_with_wrapping, the linesinfo.cursor y position was
already in relative screen space, which was simply not taken into
account.

This however reveals another bug in adjust_view: the dragging code when
scrolling down is incorrect and causes the cursor to fall off the
screen.

Also refactors a bit the cursor access code in linesinfo.