New implementation for token based movement
This new implementation relies on a function abstraction for finding the
next token in a line, given the previous token. This should allow the
movement patterns to be appplied to a lot more cases, and notably allows
one to match more complex things like 0xdeedbeef or 4.56L, or even plug
in regexes or something else.

The code is ugly is should be simplified. It uses options a lot and
could be improved with the standard option functionals like map, fmap,
and get_or_default.

At the moment:
 go_start, go_end works
 go_left, go_right mostly works but can skip complete number of blocks
 go_up and go_down do no work