Fix put_color_rect crashes
instead of asserting or trying to skip rectangles, I realized the most
practical thing to do was to reclip rectangles to the underlying window.

Given that color rectangles can come from many computations, it would be
too bothersome to double check all sources and wrap them with clipping
code or window dimension guards.