movement: experiment rewriting the block movement logic
Initially the block movement logic was based on token finding which was
line based. Now that a Cursor type has been introduced and that
TokenFinder has been rewritten in term of Cursor, it seems much more
natural to scan the line using the Cursor iterator directly.

This patch is a quick experiment for rewriting one of block movement
function directly in term of cursor.

Code looks simpler and is certainly more efficient too.

More importantly it does not refer to a full line "entity" explicitly.